requirements.txt
beautifulsoup4==4.13.5
customtkinter==5.2.2
google_api_python_client==2.116.0
google_auth_oauthlib==1.2.0
protobuf==6.32.0
Requests==2.32.5
